to checkout

+++ BackEnd/Model/ProductRating.cs
@@ -8,7 +8,7 @@ namespace BackEnd.Model
         [StringLength(3000)]
         public string Comment { get; set; }
         [Range(0, 5)]
-        public float Rating { get; set; } = 5;
+        public int Rating { get; set; } = 5;
 
         public Guid OrderItemId { get; set; }
         public OrderItem OrderItem { get; set; }
